results/rebuttal/run-20231111-034248_LLMatic-zero_shot_donnemartin-system-design-oop-finance_0_20-runs_log/finance/15/LLMatic/test_auth.py
import auth

def test_register_user():
	assert auth.register_user('test_user', 'test_password') == True
	assert auth.register_user('test_user', 'test_password') == False


def test_login_user():
	assert auth.login_user('test_user', 'test_password') == True
	assert auth.login_user('test_user', 'wrong_password') == False
	assert auth.login_user('wrong_user', 'test_password') == False


def test_logout_user():
	assert auth.logout_user('test_user') == True
	assert auth.logout_user('wrong_user') == False


def test_recover_password():
	assert auth.recover_password('test_user') == True
	assert auth.recover_password('wrong_user') == False
